addressed comment, consolidated services and fixed import related issues
