+++ byoeb-v1/byoeb/byoeb/chat_app/configuration/dependency_setup.py
@@ -75,7 +75,8 @@ from byoeb.handler import (
     UsersHandler
 )
 
-from byoeb.services.databases.mongo_db import UserMongoDBService, MessageMongoDBService
+from byoeb.services.user import UserService
+from byoeb.services.message import MessageService
 
 SINGLETON = "singleton"
 
@@ -90,11 +91,12 @@ mongo_db_factory = MongoDBFactory(
     scope=SINGLETON
 )
 
-user_db_service = UserMongoDBService(
+user_db_service = UserService(
     config=app_config,
     mongo_db_factory=mongo_db_factory
 )
-message_db_service = MessageMongoDBService(
+message_db_service = MessageService(
+    user_service=None,  # Will be set later
     config=app_config,
     mongo_db_factory=mongo_db_factory
 )
