+++ byoeb-v1/byoeb/byoeb/services/databases/mongo_db/__init__.py
@@ -1,9 +1,5 @@
 from byoeb.services.databases.mongo_db.base import BaseMongoDBService
-from byoeb.services.databases.mongo_db.user_db import UserMongoDBService
-from byoeb.services.databases.mongo_db.message_db import MessageMongoDBService
 
 __all__ = [
-    "BaseMongoDBService",
-    "UserMongoDBService",
-    "MessageMongoDBService"
+    "BaseMongoDBService"
 ]
