+++ byoeb-v1/byoeb/byoeb/services/message/message_service.py
@@ -10,13 +10,29 @@ from zoneinfo import ZoneInfo
 from byoeb.repositories.repository_factory import get_repository_factory
 from byoeb.services.user.user_service import UserService
 from byoeb.services.leaderboard.time_window_strategies import TimeWindowStrategy, TimeWindowFactory
+from byoeb.factory import MongoDBFactory
+from byoeb.services.databases.mongo_db.base import BaseMongoDBService
+from byoeb_core.models.byoeb.message_context import ByoebMessageContext
+from byoeb_integrations.databases.mongo_db.azure.async_azure_cosmos_mongo_db import AsyncAzureCosmosMongoDBCollection
+import byoeb.services.chat.constants as constants
+from byoeb.models.consensus import Consensus
 
 IST = ZoneInfo("Asia/Kolkata")
 
-class MessageService:
+class MessageService(BaseMongoDBService):
     """Service class for message-related operations."""
 
-    def __init__(self, user_service: UserService):
+    def __init__(self, user_service: UserService, config=None, mongo_db_factory: MongoDBFactory = None):
+        # Initialize BaseMongoDBService if config and factory are provided
+        if config and mongo_db_factory:
+            super().__init__(config, mongo_db_factory)
+            self.collection_name = self._config["databases"]["mongo_db"]["message_collection"]
+        else:
+            # For backward compatibility when used without MongoDB setup
+            self._config = None
+            self._mongo_db_factory = None
+            self.collection_name = None
+
         self._user_service = user_service
         self._repository_factory = None  # Initialized lazily
 
@@ -404,3 +420,215 @@ class MessageService:
                     })
 
         return results
+
+    # MongoDB-specific methods (consolidated from MessageMongoDBService)
+
+    async def get_bot_messages_by_ids(self, bot_message_ids: List[str]) -> List[ByoebMessageContext]:
+        """Fetch multiple bot messages from the database."""
+        if not self.collection_name:
+            raise ValueError("MongoDB not configured for this service instance")
+        message_collection_client = await self._get_collection_client(self.collection_name)
+        messages_obj = await message_collection_client.afetch_all({"_id": {"$in": bot_message_ids}})
+        return [ByoebMessageContext(**msg_obj["message_data"]) for msg_obj in messages_obj]
+
+    async def get_bot_messages_by_status(self, status: str) -> List[ByoebMessageContext]:
+        """Fetch bot messages with the given status."""
+        if not self.collection_name:
+            raise ValueError("MongoDB not configured for this service instance")
+        message_collection_client = await self._get_collection_client(self.collection_name)
+        messages_obj = await message_collection_client.afetch_all({"message_data.message_context.additional_info.status": status})
+        return [ByoebMessageContext(**msg_obj["message_data"]) for msg_obj in messages_obj]
+
+    async def get_latest_bot_messages_by_timestamp(self, timestamp: str):
+        """Fetch bot messages with timestamps greater than the given timestamp."""
+        if not self.collection_name:
+            raise ValueError("MongoDB not configured for this service instance")
+        message_collection_client = await self._get_collection_client(self.collection_name)
+        messages_obj = sorted(
+            await message_collection_client.afetch_all({"timestamp": {"$gt": timestamp}}),
+            key=lambda msg: msg["timestamp"],
+            reverse=True  # Sorting in descending order
+        )
+        return [ByoebMessageContext(**msg_obj["message_data"]) for msg_obj in messages_obj]
+
+    def correction_update_query(
+        self,
+        byoeb_user_messages: List[ByoebMessageContext],
+        byoeb_expert_message: ByoebMessageContext
+    ):
+        if byoeb_expert_message.reply_context is None:
+            return []
+        for byoeb_user_message in byoeb_user_messages:
+            reply_context = byoeb_user_message.reply_context
+            update_id = reply_context.additional_info.get(constants.UPDATE_ID)
+            reply_context.reply_id = update_id
+            byoeb_user_message.reply_context = reply_context
+        expert_modified_timestamp = byoeb_expert_message.reply_context.additional_info.get(constants.MODIFIED_TIMESTAMP)
+        update_data = {
+            "$set":{
+                "message_data.message_context.additional_info.correction_en_text": byoeb_expert_message.reply_context.additional_info.get(constants.CORRECTION_EN),
+                "message_data.message_context.additional_info.correction_source_text": byoeb_expert_message.reply_context.additional_info.get(constants.CORRECTION_SOURCE),
+                "message_data.message_context.additional_info.modified_timestamp": expert_modified_timestamp,
+            }
+        }
+        expert_update_queries = [({"_id": byoeb_expert_message.reply_context.reply_id}, update_data)]
+        user_update_queries = []
+        for byoeb_user_message in byoeb_user_messages:
+            update_data = {
+                "$set":{
+                    "message_data.message_context.additional_info.corrected_en_text": byoeb_user_message.message_context.message_english_text,
+                    "message_data.message_context.additional_info.corrected_source_text": byoeb_user_message.message_context.message_source_text
+                }
+            }
+            user_update_queries.append(({"_id": byoeb_user_message.reply_context.reply_id}, update_data))
+        return expert_update_queries + user_update_queries
+
+    def idk_status_update_query(
+        self,
+        byoeb_message: ByoebMessageContext,
+    ):
+        message_id = byoeb_message.message_context.message_id
+        status = byoeb_message.message_context.additional_info.get(constants.STATUS, None)
+        consensus_answer_en = ""
+        consensus_answer_source = ""
+        if status is None:
+            return
+        if status == constants.RESOLVED:
+            consensus_answer_en = byoeb_message.message_context.additional_info.get(constants.CONSENSUS_ANSWER_EN)
+            consensus_answer_source = byoeb_message.message_context.additional_info.get(constants.CONSENSUS_ANSWER_SOURCE)
+        update_data = {
+            "$set":{
+                f"message_data.message_context.additional_info.{constants.STATUS}": status,
+                f"message_data.message_context.additional_info.{constants.CONSENSUS_ANSWER_EN}": consensus_answer_en,
+                f"message_data.message_context.additional_info.{constants.CONSENSUS_ANSWER_SOURCE}": consensus_answer_source,
+                f"message_data.message_context.additional_info.{constants.MODIFIED_TIMESTAMP}": str(int(datetime.now(timezone.utc).timestamp()))
+            }
+        }
+        return ({"_id": message_id}, update_data)
+
+    def audio_idk_status_update_query(
+        self,
+        byoeb_user_message: ByoebMessageContext,
+    ):
+        message_id = byoeb_user_message.reply_context.additional_info.get(constants.BOT_AUDIO_IDK_MESSAGE_ID)
+        print("message_id", message_id)
+        status = byoeb_user_message.reply_context.additional_info.get(constants.STATUS, None)
+        print("status", status)
+        if status is None:
+            return []
+        update_data = {
+            "$set":{
+                f"message_data.message_context.additional_info.{constants.STATUS}": status,
+                f"message_data.message_context.additional_info.{constants.MODIFIED_TIMESTAMP}": str(int(datetime.now(timezone.utc).timestamp()))
+            }
+        }
+        return [({"_id": message_id}, update_data)]
+
+    def verification_status_update_query(
+        self,
+        byoeb_user_messages: List[ByoebMessageContext],
+        byoeb_expert_message: ByoebMessageContext
+    ):
+        for byoeb_user_message in byoeb_user_messages:
+            reply_context = byoeb_user_message.reply_context
+            update_id = reply_context.additional_info.get(constants.UPDATE_ID)
+            reply_context.reply_id = update_id
+            byoeb_user_message.reply_context = reply_context
+        verification_status_param = constants.VERIFICATION_STATUS
+        expert_verification_status = byoeb_expert_message.reply_context.additional_info.get(verification_status_param)
+        expert_modified_timestamp = byoeb_expert_message.reply_context.additional_info.get(constants.MODIFIED_TIMESTAMP)
+        user_verification_status = byoeb_user_messages[0].reply_context.additional_info.get(verification_status_param)
+        user_modified_timestamp = byoeb_user_messages[0].reply_context.additional_info.get(constants.MODIFIED_TIMESTAMP)
+        update_data = {
+            "$set":{
+                "message_data.message_context.additional_info.verification_status": expert_verification_status,
+                "message_data.message_context.additional_info.modified_timestamp": expert_modified_timestamp,
+                "message_data.cross_conversation_context.messages_context.$[].message_context.additional_info.verification_status": user_verification_status
+            }
+        }
+        expert_update_queries = [({"_id": byoeb_expert_message.reply_context.reply_id}, update_data)]
+        user_update_queries = []
+        for byoeb_user_message in byoeb_user_messages:
+            update_data = {
+                "$set":{
+                    "message_data.message_context.additional_info.verification_status": user_verification_status,
+                    "message_data.message_context.additional_info.modified_timestamp": user_modified_timestamp
+                }
+            }
+            user_update_queries.append(({"_id": byoeb_user_message.reply_context.reply_id}, update_data))
+        return expert_update_queries + user_update_queries
+
+    def consensus_update_query(self, user_message: ByoebMessageContext, cross_convs: List[ByoebMessageContext]):
+        consensus_list = user_message.message_context.additional_info.get(constants.CONSENSUS)
+        if consensus_list is None:
+            consensus_list = []
+        for cross_conv in cross_convs:
+            consensus = Consensus(
+                user_id = cross_conv.user.user_id,
+                status = constants.WAITING,
+                message_id = cross_conv.message_context.message_id,
+            )
+            consensus_list.append(consensus.model_dump())
+        update_data = {
+            "$set":{
+                "message_data.message_context.additional_info.consensus": consensus_list
+            }
+        }
+        return [({"_id": user_message.message_context.message_id}, update_data)]
+
+    def message_create_queries(self, byoeb_messages: List[ByoebMessageContext]) -> List[Dict[str, Any]]:
+        """Generate create queries for messages."""
+        if not byoeb_messages:
+            return []
+        return [
+            {
+                "_id": message.message_context.message_id,
+                "message_data": message.model_dump(),
+                "timestamp": str(int(datetime.now(timezone.utc).timestamp())),
+            }
+            for message in byoeb_messages
+        ]
+
+    def aggregate_queries(
+        self,
+        results: List[Dict[str, Any]]
+    ):
+        new_message_queries = {
+            constants.CREATE: [],
+            constants.UPDATE: [],
+        }
+        for queries, _, err in results:
+            if err is not None or queries is None:
+                continue
+            message_queries = queries.get(constants.MESSAGE_DB_QUERIES, {})
+            if message_queries is not None and message_queries != {}:
+                message_create_queries = message_queries.get(constants.CREATE,[])
+                message_update_queries = message_queries.get(constants.UPDATE,[])
+                new_message_queries[constants.CREATE].extend(message_create_queries)
+                new_message_queries[constants.UPDATE].extend(message_update_queries)
+
+        return new_message_queries
+
+    async def execute_queries(self, queries: Dict[str, Any]):
+        """Execute message database queries."""
+        if not queries or not self.collection_name:
+            return
+
+        message_client = await self._get_collection_client(self.collection_name)
+        if queries.get("create"):
+            await message_client.ainsert(queries["create"])
+        if queries.get("update"):
+            await message_client.aupdate(bulk_queries=queries["update"])
+
+    async def delete_message_collection(self):
+        """Delete the message collection."""
+        if not self.collection_name:
+            raise ValueError("MongoDB not configured for this service instance")
+        try:
+            message_client = await self._get_collection_client(self.collection_name)
+            if isinstance(message_client, AsyncAzureCosmosMongoDBCollection):
+                await message_client.adelete_collection()
+                return True, None
+            return False, None
+        except Exception as e:
+            return False, e
